% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ams_ic_ontology.R
\docType{import}
\name{reexports}
\alias{reexports}
\alias{read_cdm_concept_triples}
\title{Objects exported from other packages}
\keyword{internal}
\description{
These objects are imported from other packages. Follow the links
below to see their documentation.

\describe{
  \item{cdmSemantic}{\code{\link[cdmSemantic]{read_cdm_concept_triples}}}
}}

